requirements.txt
playwright
python-dotenv 
requests
tqdm
asyncio

